Remove generic params from cache methods
It doesn't work to have a generic parameter on the method without also
providing that parameter to the method some how (i.e. as a class
parameter).

So in most cases it makes much more sense to simply accept Objects. In
the case of providing a ValueLoader, we can guarantee the return type.