Add a system property for suppressing the console
Note that this change does not enforce compliance, but merely exposes
a property constant for use in downstream console implementations.

Some use cases produces a mountainous torrent of output to
System.(err|out), and the graphical console cannot keep up.

While I would argue that these use cases should be altered to use the
LogService at appropriate logging levels, it is still nice to have a
way to disable the graphical console to sidestep such problems.

However, we don't want to make it too easy for naive users to disable
the graphical console -- otherwise we will be back to where we started
with people reporting "silent failure" bugs because the associated
output was suppressed.