Add Java-side support for -D arguments
In most cases, there is no need to parse these on the native side;
Java can handle it and set the relevant system property.

It is only an issue if the system property needs to be set before
the JVM starts up for some reason. There are not too many of these
(though java.library.path comes to mind...).