Fix issues with Readers of ScriptInfo used multiple times without reset
ScriptModule, for example, does not reset the Reader aquired via
`ScriptInfo.getReader()`. While it is definitely an option to fix this
issue by resetting it there, this way we can ensure that the behaviour of
`ScriptInfo.getReader()` stays consistent. Also, this solution makes way
to some day maybe run ScriptModules in parallel.

Signed-off-by: Squareys <Squareys@googlemail.com>