DefaultPrefService: wrap all Preferences access
This encapsulates direct usage of the java.util.prefs API to:

1) a SmartPrefs class with many of the same methods (and a few new ones)
2) four remaining private helper methods of DefaultPrefService itself.

We no longer explicitly import java.util.prefs at the top of the class,
to make it easier to verify that the _only_ usages are as stated above.

The SmartPrefs class wraps a java.util.prefs.Preferences object,
providing much of the same API, as well as some new convenience API
matching that of the PrefService itself (notably: collections).

This change paves the way for more fine-grained control over our exact
usage of the Preferences API. In particular, see bug-fix in next commit.