DefaultPrefService: fix bug with keys >80 chars
The Java Preferences API does not allow keys greater than 80 characters
long (the Preferences.MAX_KEY_LENGTH constant).

As a workaround, we limit the key length internally, using the last 77
characters of the string prepending with "..." to suggest this occurred.

This means, of course, that two keys with identical latter 77 characters
will now stomp each other. But there is little we can do aside from
using hashes or some such, which would mangle the keys too much IMHO
(configuration files on disk would be incomprehensible, for example).

The fix is also incomplete when it comes to persisting maps: if you
have a map with such long keys, they will be mangled when restored.
But I do not have enough time and energy to address that right now.

Tripped over by Kyle Harrington.