Context: allow multi-injection in non-strict mode
With scijava.context.strict set to false (i.e.: strict=false flag of the
Context), injecting an object multiple times is now allowed.

This avoids the dreaded "Context already injected" exception which is
thrown, for example, when reusing an op with a second module.

This commit is dedicated to Christian Dietz!