DefaultScriptInterpreter: improve constructors
The old constructor is deprecated.

There is now a simple constructor taking only the ScriptLanguage,
as well as one which allows to manually override the ScriptEngine.

The SciJava application context is derived from the ScriptLanguage,
and services (notably: the PrefService) are injected from there.

The PrefService is now optional, and when it is null, there is no
history available. (The code for null History was already in place
everywhere -- it just couldn't happen previously due to the explicit
"new History" call in the constructor. So now we make use of it.)