ScriptInterpreter: allow multi-line evaluation
This adds new methods designed to support evaluation of multi-line
statements which are fed to the interpreter one line at a time. The
approach works according to a strategy implemented by Jason Sachs on
StackOverflow at:

    http://stackoverflow.com/a/5598207

This functionality will be useful for improving the behavior of
REPL (Read-Eval-Print-Loop) shells backed by this interpreter.