DefaultScriptService: fix null context bug
For languages created simply by wrapping an existing ScriptEngineFactory
from the ScriptEngineManager, no context was ever being injected into
the wrapped ScriptLanguage. The ScriptLanguage plugin type extends
RichPlugin, which means every plugin instance is supposed to have a
non-null context. Hence, this behavior should be considered a bug.

Since the ScriptLanguageIndex is context-free, it does not know the
context to inject it. So to fix this problem in the least invasive way,
we simply have the ScriptService walk the whole list of available
ScriptLanguages after initially populating them, injecting the context
into any language for which it is still null.