ScriptService: getScript now creates the script
To avoid a proliferation of case logic and repeated creation code,
ScriptService needs public API that retrieves a cached script and
creates it if it doesn't exist.

It is not necessary to have both a lookup-only and a lookup-and-create
method: if an API consumer wants to check the existence of a ScriptInfo,
the containing Collection is accessible.

Thus getScript now simply creates the ScriptInfo if it does not already
exist.