ConsoleUtils: attempt to match singleton params
It is not practical to require users to know the names of script
parameters to call the script.

If script params are mapped with =, great. If not, we can try to match
them with the parameters of the script.

NB: this requires that script parameters be written in a reasonable
ordering, such that user-supplied params come before system-supplied
params.