DisplayPostprocessor: set name if creating display
The Display name is not guaranteed to be null by default: it is up to
the implementation of each DisplayPlugin. A common case is that the
default name will come back as "Untitled-#" if not specified, which will
then supercede the requested output name.

To avoid this, the DisplayPostprocessor should always pass the default
name when creating a display.