ModuleService: getSingleIn/Output for classes
Add getSingleInput and getSingleOutput methods that can take a
collection of classes to match. If there is a single unresolved input
that matches any of these classes it will be identified.

This enables generalized preprocessors that may want to handle multiple
classes with similar logic (e.g. classes that can be easily converted).