ClassUtils: throw exceptions in loadClass methods
We keep the old behavior of returning null, both for convenience and for
backwards compatibility. But we add a "boolean quietly" flag which, when
set to false, makes the methods throw IllegalArgumentException instead.

This avoids problems where downstream code needs to know _why_ a class
couldn't be loaded, but the exception was eaten in the loadClass method.