When class loading fails, chain the reason
Several places in SciJava Common call the ClassUtils.loadClass routine.
Previously, that routine simply returned null without explanation when
the class could not be loaded for any reason. Now, by setting the
'quietly' flag to false, we can do proper exception handling so that the
reason for failure is known and can be passed along to the caller.

So this commit does that in all the places where it makes sense.

This commit is dedicated to Eike Heinz!