DisplayService: add a createDisplayQuietly method
This method creates a display without publishing a DisplayCreatedEvent.
It is the caller's responsibility to deal with ramifications of that.

This method is introduced to maintain backwards compatibility with the
old behavior. The display framework will probably be substantially
redesigned soon, at which point all of this will become moot. But in the
meantime, it is handy to be able to create a display without showing it.

This commit is dedicated to Richard Domander!