Always use 'log' for LogService variable name
In the vast majority of cases, when referring to a LogService, we name
it log rather than logService. This does break the more general rule of
referring to most service instances in full lower camel case, but in the
case of the LogService, it makes the code more concise yet precise.