Remove email address from @author tags
ImageJ convention is to use only the name, without email address.
The email address can be found in the pom.xml file and/or Git history.

Philosophically, we really do not want to encourage anyone to email us
directly about the code -- public communication channels are better.
And pragmatically, putting the email address in source in an invitation
for it to get out of sync.