CommandRunner: do not add new command to the index
I may regret this later. But it is much simpler and easier to understand
if we just run the command class directly, rather than wrapping it in a
new CommandInfo and then adding it to the PluginIndex.

If we later decide to restore the PluginIndex-modifying behavior, let's
do it in a nicer way where we have a method in CommandService that does
this work, rather than needing to recapitulate its internals here.