ScriptInfo: append return value if no outputs
We want the return value to be an easy way to return a single, untyped
(Object) output. This is quite useful for the "lazy" (I mean "elegant")
script writer. But for those consuming scripts, it is annoying to always
have to deal with this extra "result" output.

As a compromise, let's only append the "result" output if:
A) no other outputs were declared; and
B) "result" was not declared as an input either.