ScriptInfo: do not mix datestamp into the version
At the time I wrote that code, I was mainly concerned with ensuring that
generated version strings do not clash. However, there are compelling
reasons to avoid "false negatives" as well (two scripts which are
actually the same, but end up with differing version strings).

When mixing in the datestamp with the hash, a false negative will happen
if the script's last modified date is not preserved when copied between
systems. This makes data provenance more difficult, because two systems
may have identical installations content-wise, but different version
strings due to timestamp skew.

So, let's keep it simpler: use the content hash if available (i.e.:
if the script contents can be read), and the datestamp only if not.