Migrate ClassUtils.getArrayClass to Types.array
This also changes the behavior of the method to throw
IllegalArgumentException, rather than returning null, if
something goes wrong with the creation of the array type.
But the deprecated method preserves the legacy behavior,
for backwards compatibility.

The new behavior is more consistent with the other methods
of the Types class -- and it is better in general for a
reusable library to do proper exception handling. If the
behavior is too bothersome, we could add "boolean quietly"
flags to every method, but it would rather clutter things.