Rename get*Service() methods of services
As we move toward SJC3, I decided to drop the "get" prefix from service
dependency accessors. This is arguably inconsistent with the rest of
SJC, and certainly it is inconsistent with the Java standard library.
However, it has some nice consequences:

* When browsing available service functionality via autocomplete,
  service accessors do not clutter the accessors method list.
* It makes chaining service accessors more streamlined and readable.

I debated dropping the "Service" suffix too, in the same manner as the
Gateways do, but decided that would be confusing at the lower level of
the API where services reside: services have other API mixed in, so a
method called e.g. object() might be confused for something which
returns relevant objects, rather than an ObjectService instance.
The Gateways do not have this issue because their primary accessors
are in fact all services, so the naming is less confusing there.