Fix race condition in alias initialization
The ServicesLoadedEvent fires asynchronously. Hence, it was
possible that the service aliases did not get added in time for
subsequent alias lookups. This was most likely in the unit tests,
where a context is spun up and then immediately queried for
aliases, before things had time to settle.

Of course, it was totally bogus to initialize the service classes
this way. We now employ a better, more consistent solution:
rather than querying the ServiceIndex, we ask the PluginIndex for
all PluginInfo<Service> objects and load their classes. This is
fine because we do not need actual Service instances to exist yet
-- we only need the Class objects to create the aliases.

This fixes a long-standing race condition which caused the
SciJava Common Jenkins build to occasionally fail.