Add default behavior for service accessors
It is easy and reasonable to fall back to asking the context directly
for the service. However, there is a downside: services which do not
override these methods to return their injected field values may have
subtle bugs relating to service dependencies and initialization order.

For example, InputService has a method EventService, which is intended
to never return null. In other words: it is intended that InputService
implementations (at least by default) _depend_ on the EventService.
If you write:

    Context ctx = new Context(InputService.class);

When DefaultInputService is on the classpath, it will be instantiated,
its @Parameter EventService field will be noted, and the
highest-priority concrete EventService implementation (i.e.,
DefaultEventService) will then be recursively instantiated and
initialized, and so forth. So you will end up with a Context containing
a DefaultInputService _and_ all of its service dependencies.

But if DefaultInputService neglects to declare an @Parameter
EventService, the way the service loader is coded right now, the
EventService dependency will not be noticed, and the Context will
contain only a DefaultInputService, whose eventService() method will
return null, resulting in NPE when calling other methods which use it.

There is one vital reason to define these default method behaviors
anyway, though: to avoid breaking SPI compatibility with downstream
components. For example, the net.imagej:imagej-legacy component
implements a LegacyConsoleService (ConsoleService is a
SingletonService), which was forced to implement the new
objectService() method when upgrading its version of scijava-common.
Unfortunately, the ImageJ Updater (net.imagej:imagej-updater) has a
serious design flaw whereby it updates itself and its dependencies
(including scijava-common) first, without updating any downstream
libraries (e.g., imagej-legacy). So we end up with a situation where
scijava-common is at 2.58.2, after the addition of the newly required
objectService() method, while imagej-legacy remains at its old version
which does not implement that method. And the Context fails to spin up.
Adding these default method behaviors is a simple way around the issue.

We should consider whether to remove these default implementations in
SJC3. And we should definitely consider how best to fix SciJava Common
so that downstream JAR skew cannot taint the Context startup so easily.
But for now, it is safest to provide these methods, in case of JAR skew.