ScriptFinder: quelch unsupported script message
If you have a script directory prefix with a mixture of scripts and
non-scripts such as JAR files (which is what Fiji has, for example),
the ScriptFinder was issuing a warning message for each and every
JAR file. This is normal and expected, so let's not warn about it,
but rather mention it only when debug logging is enabled.