Make the HeadlessUI a real boy
The HeadlessUI is now a SciJava UserInterface plugin, just like
any other user interface. This is useful because while we want it
to always be chosen when headless mode is active, it should also
be used in non-headless mode when nothing else is available.
This will help prevent NPEs etc. in UIService method calls.

Consequently, this commit adds some defensive checks to
DefaultUIService. These are needed now with the new HeadlessUI,
because that UI returns null for methods like dialogPrompt --
something that no other UIs did previously. In other words: we
relaxed the postconditions of some UserInterface methods to allow
for null returns, and therefore the DefaultUIService needs to
behave accordingly, handling those nulls.