StderrLogService: use stdout non-severe messages
The logging page at http://imagej.net/Logging states that severe logging
messages (WARN and ERROR levels) are emitted to stderr, while less
severe ones (INFO and DEBUG levels) are emitted to stdout.

This was not actually true with the default LogService implementation.
This commit addresses the situation.

Originally, I planned to make this change with SciJava Common 3.0.0,
because I thought we would need to break backwards compatibility to
achieve it. But actually, the change was easy to make in a smooth
manner. SJC3 will see a cleanup of the logging API, but that cleanup
is not a prerequisite to fixing the situation with stdout + stderr.

Thanks to Florian Jug for spurring this change.