RunArgument: fix bug when argument list is given
When an argument list is provided, we need to remove it from the list
of pending command line arguments to process.

Otherwise, it will get "double" processed, which can result in
various weirdness. For instance, in ImageJ2 with the legacy layer
active, the class net.imagej.legacy.LegacyCommandline$Filename.handle
would try to handle the empty string as a filename and then barf.