Make Module support Initializable
It might be easier for the delegate class inside the module to implement
this interface than to specify the initializer inside the Plugin
annotation, especially because annotation could not be partially
overridden. Now the Initializable interface has higher priority than the
initializer annotation, and at most one of them will be called when a
module is initialized.

Closes #249.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>