Create Initializable
This interface is moved from imagej-ops. It might be easier for the
underlying class inside the module to implement this interface than to
specify the initializer inside the Plugin annotation, especially because
annotation could not be partially overridden. Now the Initializable
interface has higher priority than the initializer annotation, and at
most one of them will be called when a module is initialized.