ScriptREPL: catch exceptions for built-in commands
Previously, we only caught exceptions when evaluating lines of script.
But the built-ins can go wrong, too; e.g., when asking a script engine
to do some operation which it does not support.

So let's just always catch all exceptions when talking to the script
engines, to make a best effort to recover nicely whenever possible.

Note that this patch is mostly indentation changes;
use "git diff -b" to see the true scope of the changes more easily.