Add a rudimentary input validation mechanism
Now, if you want to validate that the value of an input is
acceptable, you can specify a validater method which will be
called when that input is marked as resolved.

Admittedly, this is a bit late in the module execution process,
and dynamic validation would still be nice in the future (e.g., to
provide feedback as values are updated in input harvester UIs).
But the good thing about this scheme is that regardless of which
preprocessor locks in the parameter value, validation will be
called immediately afterward, before the next preprocessor runs.

Thanks to Richard Domander for the idea and discussion.