Do not inject delegate command classes
Instead, let the ServicePreprocessor handle them.

It turns out it is very tricky to inject the context into the Command
instance at the proper time. The best time would be precisely when the
CommandModule instance has its context injected, but unfortunately,
there is no hook to trigger behavior in response to context injection.
It is not guaranteed that Contextual#setContext(Context) be called.
Injecting the context into the Command instance during
ContextModule#initialize() appeared to work on the surface, but actually
deferred population of Service parameters until the InitPreprocessor
was called; such behavior was unintuitive and dare I say bogus.

Furthermore, by suppressing Service and Context parameters from the list
of Command inputs, we made commands behave differently than scripts: the
latter report Service and Context parameters as input parameters,
whereas the former did not. With this change, that discrepancy is fixed.