Add support for multiple logging channels
We sometimes need Logger instances other than the LogService itself.
All the logic of AbstractLogService is hereby migrated to a separate
DefaultLogger class, instances of which are created on demand in
response to LogService#channel(String) calls. Created Logger instances
are stored in a hash by name.

The AbstractLogService maintains a special instance of DefaultLogger
as a channel called "default", which receives the log events fed
to the LogService itself. Alternately, consuming code can call
LogService#channel(String) to obtain a different channel to which
information can be logged.