Context: recursively inject the context
For each non-Service, non-Context, non-primitive @Parameter, check
whether it is already assigned a value, and if so, inject the context
into that value.

This makes it much easier to compose a class whose member fields also
need to know the context at the same time as the containing class.

See next commit for an example.