ConsoleService: guard against infinite loops
The design of the ConsoleArgument handlers allows each handler to remove
one or more arguments from the beginning of the linked argument list,
which provides for a nice sequential processing sequence.

However, it is very easy to code a ConsoleArgument and then forget to
remove any elements from the list. Previously, such a bug would cause
the ConsoleService's processArgs method to never return, instead calling
the same handler over and over.

This change adds a rudimentary safeguard against such an occurrence: if
the list of arguments is totally unchanged after a ConsoleArgument
plugin has supposedly handled it, then we warn the user about the
improper handling, remove the first argument from the list, and proceed.
In this way, infinite loops should be a lot less likely; a bogus
ConsoleArgument plugin would now need to mutate the list (e.g., add an
argument rather than remove one) in order to cause trouble.