ModuleService: be smart about maps passed to run
This hack works around an issue where some script languages,
notably Jython but potentially others too, invoke the wrong run
method when called with a map. The Object... varargs method is
chosen instead of the Map method, with the map being passed as
the sole element of the object array. The code below detects
this situation, propagating the map entries into the new map.

Thanks to Robert Haase for the suggestion.

Closes scijava/scripting-jython#8.