HeadlessArgument: move to the correct package
It was in the wrong package. The org.scijava.console package is lower
level than org.scijava.ui. Logic relating to the union of those two
packages goes into org.scijava.ui.console.

This is technically a breaking API change, but in this case we are
moving a plugin implementation class, which no one should be relying on
directly. It is probably worth a formal rule stating that SJC public API
(covered by SemVer) does not include `@Plugin` implementations at all.