LogService: provide factory methods for private Loggers
A client can retrieve a private logger by calling Loggers.newRoot().
This ListenableLogger will be independent of LogService.

If the logging of a client should be disable Loggers.silent() can be
used as Logger.

NB: DefaultListenableLogger is not public because it's implementation
might change for maintenance reasons, and client code should not depend
on it.