ConsoleService: replace MultiplePrintStream
Besides being hard to understand, the initialization of
MultiplePrintStream in DefaultConsoleService was not synchronized
correctly. The global resources System.out and System.err where
modified like this: read, modify, write. This can fail if two
DefaultConsoleServices (in two scijava contexts) do it in parallel.

This problem is solved in ListenableSystemStreams by using singletons.