ThreadService: add ability to queue jobs flexibly
Previously, you could only queue a job to the special dispatch thread
(typically the EDT). But in some scenarios, the EDT is exactly the wrong
one to use. This commit adds the capability to queue jobs to any
arbitrarily named queue. Essentially, each ID corresponds to a different
single thread processing its jobs one at a time.

This commit is dedicated to Hadrien Mary.