Allow script processors to modify the script
In particular, we want the #@ processing directives to disappear from
the executed version of the script, since they are invalid syntax for
many of the script languages.

This change gives ScriptProcessor plugins the flexibility to modify each
line of the script however they choose, although our initial use case
here will simply blank out lines which have been handled.

What processors should not do is change the number of lines; we want the
line numbers in error messages to match those of the original script.