WIP: add support for unpacking compressed files
Gotta think carefully about it.

It probably only makes sense to unpack a compressed file into a local cache.
We only want to do it if the timestamp or checksum indicates we need to.
We cannot compare lengths.

Relatedly, do we want the ability to download to a null destination --
i.e., _only_ cache? In that case, we probably want the cached
destination returned in the Download::getDestination method, no?

Can we be automagical and always unpack in the given destination is
local directory? Too magical? Cases which don't work with that scheme?