Remove unnecessary code from DefaultConverter
This also fixes the corresponding test. Since the NullConverter now
handles both null src and dest, the null logic was unnecessary.
Additionally, the casting logic was unnecessary due to the
CastingConverter. The array and number logic is still necessary, since
there's no converters for List to arrays and the number converters only
handle widening conversions.