Rename LogServiceTest to StderrLogServiceTest
It specifically tests the StderrLogService. The StderrLogService is the
de facto DefaultLogService at the moment, and the convention in other
places is for Default<Foo> to have a corresponding <Foo>Test rather than
a Default<Foo>Test. But we'll (perhaps temporarily) break the rule here
to make the git history cleaner for upcoming commits.