Remove Listenable interface
After discussion with @maarzt, we agreed that we do not need this
interface. Most things use the event bus. The only exceptions are
LogService and ConsoleService, because we want to maximize performance
of event notification. The Listenable interface has the property that it
cannot be implemented more than once, which in this case could
theoretically be an issue if you want to support notification of
multiple event types. While this would not be a problem initially,
there is also no great need to keep this interface so general.