SaveInputsPreprocessor: fix priority declaration
Technically, it worked fine as written because the declared priority
matched the PRIORITY constant. But it is more correct for the class to
use its own priority constant explicitly. In case that constant value
ever changes (although changing it would break backwards compatibility
due to the Java compiler inlining constants), we want to avoid any
chance of skew between the constant and the actually declared priority.