ScriptInfo: improve behavior of script identifiers
Script IDs are now of the form "script:${name}" for named scripts,
"script:${path}" for unnamed scripts with a file path given, and
"script:<${hex}>" where "${hex}" is a hexidecimal hash code of the
script contents when it has neither name nor path.

In this way, script IDs should hopefully always be unique, except
in the case where they are intended to collide -- e.g., to
dynamically replace one version of a script with another.