PrefService: make String arg always name, not key
Passing a raw key is never what we want. The name needs to be
synthesized with the class name. Otherwise the raw key is stored in the
_package_ node, and not actually associated with the specified class.

This is a quasi-breaking change, in that certain kinds of persisted
preferences from previous will be lost. But the behavior was confusing
and inconsistent and problematic as it was, and needed to be fixed.