CommandService: don't run Command/Module hybrids
Unfortunately, running such a command (e.g., DynamicCommand) results in
the command instance -- and not a CommandModule -- being returned by
ModuleService#run. Rather than fix the API and break backwards
compatibility, we apply a band-aid for now: this commit checks for that
situation and proactively throws an IllegalArgumentException, rather
than throwing a cryptic ClassCastException after the fact.

This commit is also dedicated to Richard Domander.