PrefService: eliminate get/putIterable methods
These have now been combined with the get/putList methods:

* The putList method accepts any iterable.
* Any persisted Iterable is retrievable as a List with getList.

Previously, it might have broken backwards compatibility for
the putList(Class, List, String) signature to be widened to
putList(Class, Iterable, String). But since we deprecated
that method in favor of a new and as-yet-unreleased method
put(Class, String, List), we can widen it to
put(Class, String, Iterable) without guilt.