PrefService: deprecate clear(Class, String)
This is the same thing as remove(Class, String).

Actually, in the Preferences API, clear and removeNode are not the same:
removeNode is recursive, whereas clear just clears keys at the current
node and not below. But the latter non-recursive behavior is not
something we want to expose via the PrefService API.