AbstractUserInterface: register viewer first
At the conclusion of DisplayViewer.view(UserInterface, Display),
a Display.update() fires, which triggers a DisplayUpdatedEvent. Since
the DisplayViewer.view happens on the EDT (in the Swing case), the
Display.update() also happens on the EDT, the DisplayUpdatedEvent is
published on the EDT immediately (via publishNow, which simply calls the
code directly). Therefore, the DisplayUpdatedEvent happens _before_ any
subsequent code after the DisplayViewer.view. So we need the new
DisplayViewer to have already been registered with the service.
Otherwise, the update will have no effect, because it does a lookup of a
DisplayViewer for the the Display in question, which will still be null.