Add convenient way to build input-only commands
The idea is to build up the inputs with code dynamically in Java, run
the command (which does nothing when executed) to exploit the module
preprocessing chain, and then harvest the final input values.

In this way, the user can be prompted for inputs dynamically from Java
code, roughly similar to the ij.gui.GenericDialog class of ImageJ 1.x.